New CipherInputStream for EXTREMELY improved decryption performance
The original Java CipherInputStream uses a 512 Byte buffer which is
extremely slow. Since that code is licensed under the GPL, I could not
simply modify that file.

I now have modified a CipherInputStream from Bouncy Castle which is MIT
licensed.

The decryption performances is now exactly as the encryption
performance. The performance improvement compared with the totally
unoptimized version is about factor 10!