Added listener for keyboard on unlocking vault.
Created a "PasswordListener" class for listening for both the "ok
button" and the "done" key (IME_ACTION_DONE). The original code used for
retrieving the input is unchanged, just moved into the new listener.
Further work is necessary, as this doesn't work with the
OnLongClickListener for renaming the vault yet.

Also, added some gitignore entries (vim swap files, error logs).